Chore: Refactor utility functions
