+++ tests/test_dummy.py
@@ -5,9 +5,9 @@ import sqlalchemy as sa
 
 from cratedb_fivetran_destination import __version__
 from cratedb_fivetran_destination.engine import Processor
-from cratedb_fivetran_destination.main import setup_logging
 from cratedb_fivetran_destination.model import TableInfo
 from cratedb_fivetran_destination.sdk_pb2 import common_pb2
+from cratedb_fivetran_destination.util import setup_logging
 
 
 def test_dummy():
